fix Playwright CT duplicate identifier error and remove redundant cross-browser testing
Extract test data to separate file to avoid Playwright CT transform conflict
when importing both a component and a value from the same module. Remove
Firefox and WebKit projects since component tests don't benefit from
cross-browser runs.
