+++ playwright-ct.config.ts
@@ -65,11 +65,7 @@ export default defineConfig({
             },
         },
     },
-    projects: [
-        { name: 'chromium', use: { ...devices['Desktop Chrome'] } },
-        { name: 'firefox', use: { ...devices['Desktop Firefox'] } },
-        { name: 'webkit', use: { ...devices['Desktop Safari'] } },
-    ],
+    projects: [{ name: 'chromium', use: { ...devices['Desktop Chrome'] } }],
     reporter: [
         ['list'],
         [
